// frontend/angular_employees/src/app/models/employee.model.ts
export class Employee {
  id?: number;
  email?: string;
  description?: string;
  published?: boolean;
  first_last_name?: any;
  second_last_name?: string;
  first_name?: string;
  other_names?: string;
  country?: string;
  identification_type?: string;
  identification_number?: string;
  hire_date?: any;
  area?: string;
}